Remove demo components for Badge and Dropdown
